backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic[email]
cryptography
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
multipart
bcrypt==4.0.1